About, Education, Profile, Recommendations: moved files to /src/components/about; index: refactors import paths accordingly
