+++ src/components/about/Recommendations.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { WebLink } from "./common/WebLink";
+import { WebLink } from "../common/WebLink";
 
 export const Recommendations = () => {
   return (
